backend/pyproject.toml
[project]
name = "shadow-role-backend"
version = "0.1.0"
description = "API RESTful for Shadow role web app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all,standard]>=0.120.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.38.0",
]
